# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Royvento event management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/eligibility/vendor/{vendorId}:
    get:
      operationId: getReviewEligibility
      tags: [reviews]
      summary: Check if the current user is eligible to leave a review for this vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewEligibility"
...
  /reviews/admin:
    get:
      operationId: listReviewsAdmin
      tags: [reviews]
      summary: List all reviews for admin moderation
      parameters:
        - name: vendorId
          in: query
          required: false
          schema:
            type: integer
        - name: rating
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: verified
          in: query
          required: false
          description: Filter by whether the reviewer has a checked-in booking at the vendor
          schema:
            type: boolean
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
  /reviews/partner:
    get:
      operationId: listReviewsPartner
      tags: [reviews]
      summary: List reviews on the authenticated partner's vendors
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
  /reviews/{reviewId}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Edit a review (owner or admin)
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReviewBody"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (owner, partner-of-vendor, or admin)
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ok"
  /reviews/event/{eventId}:
    get:
      operationId: listEventReviews
      tags: [reviews]
      summary: Reviews for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
  /reviews/vendor/{vendorId}:
    get:
      operationId: listVendorReviews
      tags: [reviews]
      summary: Reviews for a vendor
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedReviews"
...
components:
  schemas:
...
    Ok:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        eventId:
          type: integer
        vendorId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
        userName:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        userImage:
          type: string
        verifiedBooking:
          type: boolean
        vendorName:
          type: string
      required: [id, userId, vendorId, rating, comment, createdAt, userName]
    ReviewEligibility:
      type: object
      properties:
        eligible:
          type: boolean
        reason:
          type: string
          enum: [ok, not_authenticated, no_checkin, already_reviewed]
        existingReviewId:
          type: integer
      required: [eligible, reason]
    UpdateReviewBody:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        imageUrls:
          type: array
          items:
            type: string
          maxItems: 5
    PaginatedReviews:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required: [items, total, page, pageSize]
    CreateReviewBody:
      type: object
      properties:
        eventId:
          type: integer
        vendorId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        imageUrls:
          type: array
          items:
            type: string
          maxItems: 5
      required: [vendorId, rating, comment]
...